requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-mongoengine==1.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.14.0
python-dotenv==0.19.0
Werkzeug==2.0.1
